refactor: move dateToRfc3339Local to shared utility module
Move dateToRfc3339Local from logs/views/filters.tsx to lib/utils/date.ts
so other features can discover and reuse it. Keep a re-export from the
original location for backwards compatibility.
